+++ crates/stackable-certs/CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-05-08
+
 ### Changed
 
+- Bump Rust dependencies and GitHub Actions ([#782]).
 - Bump kube to 0.89.0 and update all dependencies ([#762]).
 - BREAKING: Bump k8s compilation version to `1.29`. Also bump all dependencies.
   There are some breaking changes in k8s-openapi, e.g. PVCs now have `VolumeResourceRequirements` instead of `ResourceRequirements`,
@@ -16,6 +19,7 @@ All notable changes to this project will be documented in this file.
 [#762]: https://github.com/stackabletech/operator-rs/pull/762
 [#769]: https://github.com/stackabletech/operator-rs/pull/769
 [#772]: https://github.com/stackabletech/operator-rs/pull/772
+[#782]: https://github.com/stackabletech/operator-rs/pull/782
 
 ## [0.2.0] - 2024-03-26