+++ crates/stackable-versioned/CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Improve action chain generation ([#784]).
+
+[#784](ttps://github.com/stackabletech/operator-rs/pull/784)
+
 ## [0.1.0] - 2024-05-08
 
 ### Changed
 
 - Bump Rust dependencies and GitHub Actions ([#782]).
-- Bump GitHub workflow actions ([#772]).
-- Revert `zeroize` version bump ([#772]).
 
-[#772]: https://github.com/stackabletech/operator-rs/pull/772
 [#782]: https://github.com/stackabletech/operator-rs/pull/782