+++ crates/stackable-webhook/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-05-08
+
 ### Added
 
 - Instrument `WebhookServer` with `AxumTraceLayer`, add static healthcheck without instrumentation ([#758]).
@@ -11,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Bump Rust dependencies and GitHub Actions ([#782]).
 - Bump kube to 0.89.0 and update all dependencies ([#762]).
 - BREAKING: Bump k8s compilation version to `1.29`. Also bump all dependencies.
   There are some breaking changes in k8s-openapi, e.g. PVCs now have `VolumeResourceRequirements` instead of `ResourceRequirements`,
@@ -28,6 +31,7 @@ All notable changes to this project will be documented in this file.
 [#769]: https://github.com/stackabletech/operator-rs/pull/769
 [#772]: https://github.com/stackabletech/operator-rs/pull/772
 [#778]: https://github.com/stackabletech/operator-rs/pull/778
+[#782]: https://github.com/stackabletech/operator-rs/pull/782
 
 ## [0.2.0] - 2024-03-26